+++ Runtime/Core/Reader/FormatLocation.cs
@@ -1,0 +1,6 @@
+public enum FormatLocation
+{
+    Line,
+    Speaker,
+    Option,
+}