+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changes
+## [beta-0.0.2] - 2025-05-18
+### Formatted Reader Global Formats
+- Added a global format option for the fomatted reader
+- Added the ability to toggle on and off clearing text on line read
+
 ## [beta-0.0.1] - 2025-05-12
 ### First release
 - First features added
 - Test suite works
 - Changelog exists
+